requirements.txt
matplotlib
numpy
scipy
hunspell
nltk
treetaggerwrapper
